// src/app/modules/billing/review/models/error-batch-detail.model.ts
export interface ErrorReportModel {
    batchNum: string;
    batchDetailId: number;
    alias: string;
    billingLevel: string;
    billingLevelCode: string;
    blngBr: string;
    sfxCode: string;
    rateCard: string;
    sfxIds: string;
    msaName: string;
    billingBy: string;
    consignerName: string;
    consineeName: string;
    status: string;
    message: string;
}